Move the methods from the `Stringifiers` to the `Identifiers` class.
