+++ client/src/main/java/org/spine3/base/Identifiers.java
@@ -20,12 +20,22 @@
 
 package org.spine3.base;
 
+import com.google.common.reflect.TypeToken;
 import com.google.protobuf.Any;
+import com.google.protobuf.Message;
+import com.google.protobuf.MessageOrBuilder;
+import com.google.protobuf.Timestamp;
+import com.google.protobuf.util.Timestamps;
 import org.spine3.Internal;
 
+import javax.annotation.Nullable;
+import java.util.Collection;
 import java.util.UUID;
+import java.util.regex.Pattern;
 
 import static com.google.common.base.Preconditions.checkNotNull;
+import static com.google.protobuf.TextFormat.shortDebugString;
+import static org.spine3.protobuf.AnyPacker.unpack;
 
 /**
  * Utility class for working with identifiers.
@@ -37,6 +47,14 @@ public class Identifiers {
 
     /** The suffix of ID fields. */
     public static final String ID_PROPERTY_SUFFIX = "id";
+    /** A {@code null} ID string representation. */
+    public static final String NULL_ID = "NULL";
+    /** An empty ID string representation. */
+    public static final String EMPTY_ID = "EMPTY";
+    private static final Pattern PATTERN_COLON = Pattern.compile(":");
+    private static final Pattern PATTERN_T = Pattern.compile("T");
+    private static final Pattern PATTERN_COLON_SPACE = Pattern.compile(": ");
+    private static final String EQUAL_SIGN = "=";
 
     private Identifiers() {
         // Prevent instantiation of this utility class.
@@ -136,5 +154,100 @@ public class Identifiers {
     public static <I> I getDefaultValue(Class<I> idClass) {
         checkNotNull(idClass);
         return Identifier.getDefaultValue(idClass);
+    }
+
+    /**
+     * Converts the passed timestamp to the string, which will serve as ID.
+     *
+     * @param timestamp the value to convert
+     * @return string representation of timestamp-based ID
+     */
+    public static String toIdString(Timestamp timestamp) {
+        checkNotNull(timestamp);
+        String result = Timestamps.toString(timestamp);
+        result = PATTERN_COLON.matcher(result)
+                              .replaceAll("-");
+        result = PATTERN_T.matcher(result)
+                          .replaceAll("_T");
+
+        return result;
+    }
+
+    /**
+     * Converts the passed ID value into the string representation.
+     *
+     * @param id  the value to convert
+     * @param <I> the type of the ID
+     * @return <ul>
+     *              <li>for classes implementing {@link Message} &mdash; a Json form;
+     *              <li>for {@code String}, {@code Long}, {@code Integer} &mdash;
+     *              the result of {@link Object#toString()};
+     *              <li>for {@code null} ID &mdash; the {@link #NULL_ID};
+     *              <li>if the result is empty or blank string &mdash; the {@link #EMPTY_ID}.
+     *         </ul>
+     * @throws IllegalArgumentException if the passed type isn't one of the above or
+     *                                  the passed {@link Message} instance has no fields
+     * @see StringifierRegistry
+     */
+    public static <I> String idToString(@Nullable I id) {
+        if (id == null) {
+            return NULL_ID;
+        }
+
+        final Identifier<?> identifier;
+        if (id instanceof Any) {
+            final Message unpacked = unpack((Any) id);
+            identifier = Identifier.fromMessage(unpacked);
+        } else {
+            identifier = Identifier.from(id);
+        }
+
+        final String result = identifier.toString();
+        return result;
+    }
+
+    @SuppressWarnings("unchecked") // OK to cast to String as output type of Stringifier.
+    static String idMessageToString(Message message) {
+        checkNotNull(message);
+        final String result;
+        final StringifierRegistry registry = StringifierRegistry.getInstance();
+        final Class<? extends Message> msgClass = message.getClass();
+        final TypeToken<? extends Message> msgToken = TypeToken.of(msgClass);
+        if (registry.hasStringifierFor(msgToken)) {
+            @SuppressWarnings("OptionalGetWithoutIsPresent") // OK as we check for presence above.
+            final Stringifier converter = registry.get(msgToken)
+                                                  .get();
+            result = (String) converter.convert(message);
+        } else {
+            result = convert(message);
+        }
+        return result;
+    }
+
+    private static String convert(Message message) {
+        final Collection<Object> values = message.getAllFields()
+                                                 .values();
+        final String result;
+        if (values.isEmpty()) {
+            result = EMPTY_ID;
+        } else if (values.size() == 1) {
+            final Object object = values.iterator()
+                                        .next();
+            if (object instanceof Message) {
+                result = idMessageToString((Message) object);
+            } else {
+                result = object.toString();
+            }
+        } else {
+            result = messageWithMultipleFieldsToString(message);
+        }
+        return result;
+    }
+
+    private static String messageWithMultipleFieldsToString(MessageOrBuilder message) {
+        String result = shortDebugString(message);
+        result = PATTERN_COLON_SPACE.matcher(result)
+                                    .replaceAll(EQUAL_SIGN);
+        return result;
     }
 }
