+++ client/src/test/java/org/spine3/base/StringifiersShould.java
@@ -44,10 +44,10 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.spine3.base.Identifiers.newUuid;
-import static org.spine3.base.Stringifiers.EMPTY_ID;
-import static org.spine3.base.Stringifiers.NULL_ID;
-import static org.spine3.base.Stringifiers.idToString;
-import static org.spine3.base.Stringifiers.toIdString;
+import static org.spine3.base.Identifiers.EMPTY_ID;
+import static org.spine3.base.Identifiers.NULL_ID;
+import static org.spine3.base.Identifiers.idToString;
+import static org.spine3.base.Identifiers.toIdString;
 import static org.spine3.protobuf.Timestamps2.getCurrentTime;
 import static org.spine3.protobuf.Values.newIntValue;
 import static org.spine3.protobuf.Values.newLongValue;
