+++ server/src/main/java/org/spine3/server/aggregate/AggregateCommandEndpoint.java
@@ -23,7 +23,7 @@ package org.spine3.server.aggregate;
 import com.google.protobuf.Message;
 import org.spine3.base.CommandContext;
 import org.spine3.base.CommandEnvelope;
-import org.spine3.base.Stringifiers;
+import org.spine3.base.Identifiers;
 import org.spine3.server.entity.LifecycleFlags;
 
 /**
@@ -137,7 +137,7 @@ class AggregateCommandEndpoint<I, A extends Aggregate<I, ?, ?>> {
         private void logConcurrentModification(I aggregateId,
                                                Message commandMessage,
                                                int newEventCount) {
-            final String idStr = Stringifiers.idToString(aggregateId);
+            final String idStr = Identifiers.idToString(aggregateId);
             final Class<?> aggregateClass = repository.getAggregateClass();
             AggregateRepository.log()
                .warn("Detected the concurrent modification of {} ID: {}. " +
