+++ server/src/main/java/org/spine3/server/entity/Repository.java
@@ -41,7 +41,7 @@ import java.util.Iterator;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static java.lang.String.format;
-import static org.spine3.base.Stringifiers.idToString;
+import static org.spine3.base.Identifiers.idToString;
 import static org.spine3.server.reflect.Classes.getGenericParameterType;
 import static org.spine3.util.Exceptions.unsupported;
